feat: add new planning files and enhance heuristic management for improved state evaluation. also added script to run coverage testing.

+++ src/heuristics/HeuristicsManager.h
@@ -36,7 +36,7 @@ public:
    * is not applicable or the state does not satisfy the goals and needs to be
    * discarded.
    */
-  short get_heuristic_value(State<StateRepr> &eState);
+  int get_heuristic_value(State<StateRepr> &eState);
 
   /**
    * \brief Computes and sets the heuristic value for a given state.
